+++ samples/test_azure_event_hubs_receive.py
@@ -16,7 +16,8 @@ except Exception:
     from urllib.parse import quote_plus
 
 import uamqp
-from uamqp import address, types, utils, authentication
+from uamqp import address, types, utils, authentication, MessageBodyType
+from uamqp.message import DataBody, SequenceBody, ValueBody
 
 
 def get_logger(level):
@@ -622,6 +623,99 @@ def test_event_hubs_not_receive_events_during_connection_establishment(live_even
         receive_client.close()
 
 
+def event_hubs_send_different_amqp_body_type_sync(live_eventhub_config):
+
+    uri = "sb://{}/{}".format(live_eventhub_config['hostname'], live_eventhub_config['event_hub'])
+    sas_auth = authentication.SASTokenAuth.from_shared_access_key(
+        uri, live_eventhub_config['key_name'], live_eventhub_config['access_key'])
+
+    target = "amqps://{}/{}/Partitions/0".format(live_eventhub_config['hostname'], live_eventhub_config['event_hub'])
+    send_client = uamqp.SendClient(target, auth=sas_auth, debug=False)
+
+    data_body_1 = [b'data1', b'data2']
+    data_body_message_1 = uamqp.message.Message(body=data_body_1)
+    data_body_message_1.application_properties = {'body_type': 'data_body_1'}
+    send_client.queue_message(data_body_message_1)
+
+    data_body_2 = b'data1'
+    data_body_message_2 = uamqp.message.Message(body=data_body_2, body_type=MessageBodyType.Data)
+    data_body_message_2.application_properties = {'body_type': 'data_body_2'}
+    send_client.queue_message(data_body_message_2)
+
+    value_body_1 = [b'data1', -1.23, True, {b'key': b'value'}, [1, False, 1.23, b'4']]
+    value_body_message_1 = uamqp.message.Message(body=value_body_1)
+    value_body_message_1.application_properties = {'body_type': 'value_body_1'}
+    send_client.queue_message(value_body_message_1)
+
+    value_body_2 = {b'key1': {b'sub_key': b'sub_value'}, b'key2': b'value', 3: -1.23}
+    value_body_message_2 = uamqp.message.Message(body=value_body_2, body_type=MessageBodyType.Value)
+    value_body_message_2.application_properties = {'body_type': 'value_body_2'}
+    send_client.queue_message(value_body_message_2)
+
+    sequence_body_1 = [b'data1', -1.23, True, {b'key': b'value'}, [b'a', 1.23, True]]
+    sequence_body_message_1 = uamqp.message.Message(body=sequence_body_1, body_type=MessageBodyType.Sequence)
+    sequence_body_message_1.application_properties = {'body_type': 'sequence_body_1'}
+    send_client.queue_message(sequence_body_message_1)
+
+    sequence_body_2 = [[1, 2, 3], [b'aa', b'bb', b'cc'], [True, False, True], [{b'key1': b'value'}, {b'key2': 123}]]
+    sequence_body_message_2 = uamqp.message.Message(body=sequence_body_2, body_type=MessageBodyType.Sequence)
+    sequence_body_message_2.application_properties = {'body_type': 'sequence_body_2'}
+    send_client.queue_message(sequence_body_message_2)
+
+    results = send_client.send_all_messages(close_on_done=False)
+    assert not [m for m in results if m == uamqp.constants.MessageState.SendFailed]
+
+    sas_auth = authentication.SASTokenAuth.from_shared_access_key(
+        uri, live_eventhub_config['key_name'], live_eventhub_config['access_key'])
+
+    source = "amqps://{}/{}/ConsumerGroups/{}/Partitions/{}".format(
+        live_eventhub_config['hostname'],
+        live_eventhub_config['event_hub'],
+        live_eventhub_config['consumer_group'],
+        0)
+
+    result_dic = {}
+    receive_client = uamqp.ReceiveClient(source, auth=sas_auth, timeout=5000, debug=False, prefetch=10)
+    gen = receive_client.receive_messages_iter()
+    for message in gen:
+        if message.application_properties and message.application_properties.get(b'body_type'):
+            if message.application_properties.get(b'body_type') == b'data_body_1':
+                check_list = [data for data in message.get_data()]
+                assert isinstance(message._body, DataBody)
+                assert check_list == data_body_1
+                result_dic['data_body_1'] = 1
+            elif message.application_properties.get(b'body_type') == b'data_body_2':
+                check_list = [data for data in message.get_data()]
+                assert isinstance(message._body, DataBody)
+                assert check_list == [data_body_2]
+                result_dic['data_body_2'] = 1
+            elif message.application_properties.get(b'body_type') == b'value_body_1':
+                assert message.get_data() == value_body_1
+                assert isinstance(message._body, ValueBody)
+                result_dic['value_body_1'] = 1
+            elif message.application_properties.get(b'body_type') == b'value_body_2':
+                assert message.get_data() == value_body_2
+                assert isinstance(message._body, ValueBody)
+                result_dic['value_body_2'] = 1
+            elif message.application_properties.get(b'body_type') == b'sequence_body_1':
+                check_list = [data for data in message.get_data()]
+                assert check_list == [sequence_body_1]
+                assert isinstance(message._body, SequenceBody)
+                result_dic['sequence_body_1'] = 1
+            elif message.application_properties.get(b'body_type') == b'sequence_body_2':
+                check_list = [data for data in message.get_data()]
+                assert check_list == sequence_body_2
+                assert isinstance(message._body, SequenceBody)
+                result_dic['sequence_body_2'] = 1
+
+            log.info(message.annotations.get(b'x-opt-sequence-number'))
+            log.info(str(message))
+
+    send_client.close()
+    receive_client.close()
+    assert len(result_dic) == 6
+
+
 if __name__ == '__main__':
     config = {}
     config['hostname'] = os.environ['EVENT_HUB_HOSTNAME']
@@ -630,4 +724,4 @@ if __name__ == '__main__':
     config['access_key'] = os.environ['EVENT_HUB_SAS_KEY']
     config['consumer_group'] = "$Default"
     config['partition'] = "0"
-    test_event_hubs_not_receive_events_during_connection_establishment(config)
+    event_hubs_send_different_amqp_body_type_sync(config)